Move shared guards (AuthGuard, BaseGuard, NotLoggedGuard) to src/core/guards
