+++ src/core/guards/auth.guard.ts
