+++ src/core/guards/base.guard.ts
