+++ src/modules/answer/answer.controller.ts
@@ -12,15 +12,15 @@ import {
   UseInterceptors,
 } from '@nestjs/common';
 import { AnswerService } from './answer.service';
-import { AuthGuard } from 'src/modules/auth/guards/auth.guard';
-import { CurrentUser } from 'src/modules/auth/decorators/current-user.decorator';
+import { AuthGuard } from 'src/core/guards/auth.guard';
+import { CurrentUser } from 'src/common/decorators/current-user.decorator';
 import { User, Answer } from 'src/database/entities';
 import { CreateAnswernDto } from './dto/create-answer.dto';
 import { QueryAnswernDto } from './dto/query-answer.dto';
 import { Serialize } from 'src/core/interceptors/serialize.interceptor';
 import { AnswerDto } from './dto/answer.dto';
 import { AnswerOwnerGuard } from './guards/answer-owner.guard';
-import { OwnerAnswer } from './decorators/owner-answer.decorator';
+import { OwnerAnswer } from 'src/common/decorators/owner-answer.decorator';
 import { AnswerOrQuestionOwnerGuard } from './guards/answer-or-question-owner.guard';
 import { PaginationDto } from './dto/pagination.dto';
 import { CacheInterceptor } from '@nestjs/cache-manager';
