+++ src/modules/follow/follow.controller.ts
@@ -12,8 +12,8 @@ import { FollowService } from './follow.service';
 import { Serialize } from 'src/core/interceptors/serialize.interceptor';
 import { Follow, User } from 'src/database/entities';
 import { FollowDto } from './dto/follow.dto';
-import { AuthGuard } from 'src/modules/auth/guards/auth.guard';
-import { CurrentUser } from 'src/modules/auth/decorators/current-user.decorator';
+import { AuthGuard } from 'src/core/guards/auth.guard';
+import { CurrentUser } from 'src/common/decorators/current-user.decorator';
 import { CacheInterceptor } from '@nestjs/cache-manager';
 
 @Serialize(FollowDto)
